requirements.txt
absl-py==0.4.1
asgiref==3.2.3
astor==0.7.1
astroid==2.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.4.3
dj-database-url==0.5.0
Django==2.1.7
django-ckeditor==5.8.0
django-classy-tags==0.9.0
django-cms==3.7.1
django-crispy-forms==1.7.2
django-easy-pdf==0.1.1
django-filer==1.6.0
django-formtools==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
djangorestframework==3.11.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
gast==0.2.0
grpcio==1.14.2
gunicorn==19.9.0
h5py==2.8.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
inflect==0.3.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10

lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
nltk==3.4.5
numpy==1.16.2
opencv-python==4.1.2.30
openpyxl==2.6.0
pandas==0.24.1
Pillow==5.2.0
protobuf==3.6.1
psycopg2==2.7.7
pylint==2.0.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
rake-nltk==1.0.4
reportlab==3.5.1
requests==2.22.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.5
SQLAlchemy==1.3.11
sqlparse==0.3.0

termcolor==1.1.0
tqdm==4.40.2
typed-ast==1.1.0
Unidecode==1.0.23
urllib3==1.25.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
wincertstore==0.2
wrapt==1.10.11
xhtml2pdf==0.2.2
XlsxWriter==1.0.5
